move config.py to utils
