+++ fastrunner/cmd/api.py
@@ -21,17 +21,16 @@ import sys
 
 from oslo_config import cfg
 from oslo_log import log as logging
-from oslo_reports import guru_meditation_report as gmr
 import six
 
-from fastrunner import config
+from fastrunner import utils
 from fastrunner import service
 
 CONF = cfg.CONF
 
 
 def main():
-    config.parse_args(sys.argv)
+    utils.parse_args(sys.argv)
     logging.setup(CONF, "fastrunner")
     log = logging.getLogger(__name__)
