+++ fastrunner/utils.py
@@ -1,11 +1,17 @@
+from oslo_cache import core as cache
+from oslo_config import cfg
+from oslo_db import options
+from oslo_log import log
+from oslo_middleware import cors
 from oslo_utils import importutils
+
 import six
 import functools
-
 
 import eventlet
 from oslo_context import context as common_context
 
+CONF = cfg.CONF
 
 def spawn(func, *args, **kwargs):
     """Passthrough method for eventlet.spawn.
@@ -59,3 +65,57 @@ def walk_class_hierarchy(clazz, encountered=None):
             for subsubclass in walk_class_hierarchy(subclass, encountered):
                 yield subsubclass
             yield subclass
+
+
+# NOTE(mikal): suds is used by the vmware driver, removing this will
+# cause many extraneous log lines for their tempest runs. Refer to
+# https://review.openstack.org/#/c/219225/ for details.
+_DEFAULT_LOG_LEVELS = ['amqp=WARN', 'amqplib=WARN', 'boto=WARN',
+                       'qpid=WARN', 'sqlalchemy=WARN', 'suds=INFO',
+                       'oslo_messaging=INFO', 'iso8601=WARN',
+                       'requests.packages.urllib3.connectionpool=WARN',
+                       'urllib3.connectionpool=WARN', 'websocket=WARN',
+                       'keystonemiddleware=WARN', 'routes.middleware=WARN',
+                       'stevedore=WARN', 'glanceclient=WARN']
+
+_DEFAULT_LOGGING_CONTEXT_FORMAT = ('%(asctime)s.%(msecs)03d %(process)d '
+                                   '%(levelname)s %(name)s [%(request_id)s '
+                                   '%(user_identity)s] %(instance)s'
+                                   '%(message)s')
+
+
+def parse_args(argv, default_config_files=None, configure_db=True,
+               init_rpc=True):
+    log.set_defaults(_DEFAULT_LOGGING_CONTEXT_FORMAT, _DEFAULT_LOG_LEVELS)
+    log.register_options(CONF)
+    cache.configure(CONF)
+    set_middleware_defaults()
+
+    CONF(argv[1:],
+         project='fastrunner',
+         default_config_files=default_config_files)
+
+
+
+def set_middleware_defaults():
+    """Update default configuration options for oslo.middleware."""
+    # CORS Defaults
+    # TODO(krotscheck): Update with https://review.openstack.org/#/c/285368/
+    cfg.set_defaults(cors.CORS_OPTS,
+                     allow_headers=['X-Auth-Token',
+                                    'X-Openstack-Request-Id',
+                                    'X-Identity-Status',
+                                    'X-Roles',
+                                    'X-Service-Catalog',
+                                    'X-User-Id',
+                                    'X-Tenant-Id'],
+                     expose_headers=['X-Auth-Token',
+                                     'X-Openstack-Request-Id',
+                                     'X-Subject-Token',
+                                     'X-Service-Token'],
+                     allow_methods=['GET',
+                                    'PUT',
+                                    'POST',
+                                    'DELETE',
+                                    'PATCH']
+                     )
